requirements.txt
google-adk
python-dotenv
fireworks-ai>=1.0.0a
fastapi[standard]
pydantic
uvicorn
streamlit
requests